requirements-dev.txt
# Test dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Pre-commit hooks
pre-commit>=3.6.0

# Code quality
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
